chore: simplify TypeScript configuration
Use the shared Node config with minimal local overrides and replace custom shims with published type packages where available.
