+++ package.json
@@ -73,9 +73,10 @@
     "@swc-node/register": "^1.11.1",
     "@swc/cli": "^0.8.1",
     "@swc/core": "^1.15.41",
-    "@swc/helpers": "^0.5.23",
     "@swc/jest": "^0.2.39",
+    "@types/common-tags": "^1.8.4",
     "@types/jest": "^30.0.0",
+    "@types/lodash.memoize": "^4.1.9",
     "@types/node": "^25.9.3",
     "@unts/patch-package": "^8.1.1",
     "all-contributors-cli": "^6.26.1",
